+++ cocos/3d/geom-utils/intersect.js
@@ -1,5 +1,5 @@
 import { mat3, vec3 } from '../../core/vmath';
-import distance from './distance';
+import * as distance from './distance';
 import enums from './enums';
 
 /**
